% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{number of times to resample the provided sample}

\item{x}{the original sample you wish to bootstrap}

\item{fun}{either a string for an r function or a function pointer to a user-defined function to examine the samples}

\item{alpha}{the cutoff for the tails in the returned confidence interval}

\item{...}{}
}
\value{
a list of the confidence interval, your chosen function, and the sample
}
\description{
Use bootstrapping to produce a sampling distribution for a given sample
}
\examples{
myboot2(x=c(1,2,3,4,5,6,7))

myboot2(x=c(1,2,3,4,5,6,7), alpha=0.25)
}
