% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci}
\usage{
myci(x, alpha = 0.05)
}
\arguments{
\item{x}{vector of sample data}

\item{alpha}{the desired confidence level}
}
\value{
the upper and lower boundaries of the confidence interval
}
\description{
find a confidence interval for a given sample's mean
}
\examples{
myci(c(1,2,3,4,5), alpha=0.005)
}
