% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{the desired sample size}

\item{iter}{the number of times to take a sample of size n}
}
\value{
a vector of the sums of each sample
}
\description{
A function to demonstrate the Central Limit Theorem on a uniform distribution.
}
\examples{
myclt(30, 100000)
}
