% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mygamma.R
\name{mygamma}
\alias{mygamma}
\title{mygamma}
\usage{
mygamma(shape, rate = 1, a)
}
\arguments{
\item{shape}{shape parameter for the gamma PDF that will be produced}

\item{rate}{rate parameter for the gamma PDF that will be produced}

\item{a}{the value at which you wish to know the -infinity to a probability of of this distribution}
}
\value{
a vector of the probability
}
\description{
Determine and visualize exponential densities using the gamma distribution
}
\examples{
mygamma(3, 1, 2)
x<-mygamma(1, 1, 3)
}
