% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the supplied function to apply to the parameter}

\item{x}{the data}

\item{param}{the parameter of interest}

\item{...}{}
}
\value{
a list of the index of highest likelihood, the parameter at that index, the log likelihood at that index, and the slope at that position.
}
\description{
function to graphically estimate the maximum log likelihood for a given parameter
}
\examples{
logbin=function(x,param) log(dbinom(x,prob=param,size=10))
mymaxlik(x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000),lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)
}
