% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the normal to observe}

\item{sigma}{the standard deviation of the normal to observe}
}
\value{
the probability of seeing a value from -infinity to a in a in this distribution
}
\description{
Determine and visualize the probability of seeing a value from -infinity to a in a normal distribution
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
