% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhist.R
\name{shinyhist}
\alias{shinyhist}
\title{shinyhist}
\usage{
shinyhist()
}
\description{
This app function uses bootstrapping to resample a sample
from a normal distribution with a mean of 0, and the variance is adjustable with a slider
}
\examples{
\dontrun{ shinyhist()}
}
