% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{- the number of n-samples to draw}

\item{n}{- sample size for each iteration}

\item{p}{- the probability of a 'Success'}
}
\value{

}
\description{
Title
}
